Optimize QPitchCore for higher FPS
We now maintain a cyclic buffer of samples.  Every time PortAudio calls
back, we add samples to the cyclic buffer, and immediately perform FFT
on the whole cyclic buffer, and update the GUI.  We also reduce the
PortAudio buffer size, using a nominated latency of 1/60 second.  This
helps the GUI achieve a much higher 60 FPS refresh rate.

We also fix a bug where the PortAudio stream started before the buffers
are ready.

We also changed the PortAudio sample format to paFloat32 because we use
floating point numbers for FFT anyway.