Refactor `QLogView::paintEvent`
Use floating-point coordinates wherever possible.  This is important for
supporting HiDPI displays.

Do not cache background.  We can already achieve 60 FPS on modern
devices even with a debug build.

Merge common code paths, especially the drawing of lables, by merging
the background and foreground drawing.

Fixed the incorrect translation of coordinates.  It was shifted too much
to the right.