Optimize callback buffer
We no longer specify the PortAudio buffer size and trust PortAudio to
deliver the optimial (and possibly varying) buffer sizes for the lowest
latency.

We lock the cyclic buffer using a separate mutex to reduce the
probability of race between the PortAudio callback thread and the
QPitchCore thread.

The PortAudio callback no longer discards samples when the QPitchCore
thread has not processed the samples from the previous callback, yet.
Instead, we acknowledge that the callback can deliver samples faster
than the QPitchCore thread can process, and simply enqueue the samples
into the cyclic buffer.

We add code for profiling the time/frequency of the PortAudio callback.