Coding style: Use '///' for doc comments
We change the style of doc comments to '///' before items.  This is
consistent with Rust's style which I am familar with.

Specifically, we switch away from '//!<' after the item.  Although it is
terse for short comments, it quickly becomes cumbersome when adding
non-trivial paragraphs of comments to a single item.  It is also
somewhat unfriendly to auto-formatting using clang-format.