Minor adjustment to UI files and classes.
We stopped setting `CMAKE_AUTOUIC_SEARCH_PATHS` to `ui`, but we
explicitly include `"ui/ui_xxxx.h"`.  Previously, CMake seem to be
unable to consider `*.ui` as a dependency of the `.cpp` files, and we
have to invoke `cmake --build .` multiple times to build.  With
this change, the problem goes away.

We let QPitch, QAboutDlg and QSettingsDlg instances hold a pointer to
the UI instances instead of including it as a member.  By doing this, we
no longer need to include "ui/ui_*.h" in `.h` files.

We changed member variables `_gt`, `_ab` and `_sd` to `_ui` for
consistency.

We slightly adjusted the layout of the main window, moving the frequency
diff view below the QLogView.

We also made a few minor changes to `CMakeLists.txt`.